[FIX] Fixing permutation and bootstrapping
